[codex] Implement matrix-charts 0.5.0-r3 phase 6

+++ matrix-charts/src/main/groovy/se/alipsa/matrix/charm/render/scale/ContinuousCharmScale.groovy
@@ -242,9 +242,9 @@ class ContinuousCharmScale extends CharmScale {
       return bd.toBigInteger().toString()
     }
     if (bd < BigDecimal.ONE) {
-      return String.format('%.2g', bd as double)
+      return String.format('%.2g', bd)
     }
-    String.format('%.0f', bd as double)
+    String.format('%.0f', bd)
   }
 
   private List<Object> resolveConfiguredBreaks() {
